+++ tests/test_twisted.py
@@ -9,11 +9,12 @@ import json
 import pytest
 
 from pretend import call_recorder
+from six import PY3
 from twisted.python.failure import Failure, NoCurrentExceptionError
 from twisted.python.log import ILogObserver
 
 from structlog._config import _CONFIG
-from structlog._compat import OrderedDict, StringIO, PY3
+from structlog._compat import OrderedDict, StringIO
 from structlog._loggers import ReturnLogger
 from structlog.twisted import (
     BoundLogger,